turbine: Fix and cleanup redundant metric (#9287)
The broadcast-process-shred-stats (and interrupted variant) have
duplicate fields for the number of data and coding shred. This is
probably a relic of when we rolled merkle shreds out. So, remove the
duplicate fields.

This also addresses a small issue where some shreds would be counted
in metrics for the wrong slot when we have an interrupted slot