bpf_loader: expose SBPF interpreter mode and debugger features (#8407)
* Introduces two features:
- sbpf-interpreter-mode that forces the Solana SBPF VM to work in
interpreter mode. It is only this mode from which register tracing
could be obtained. On MacOS it is this mode that is always used, but
on Linux it always defaults to using jit compiled programs. So allow
Linux users to also be able to work in interpreter mode.
- sbpf-debugger is to provide a way to enable the debugger feature
in the solana-sbpf crate. The debugger is tied to interpreter mode
and hence upon enabling this feature switch to interpreter mode.

* Refactors: simplify use_jit initialization with cfg_if!

* Remove the sbpf-interpreter-mode feature.

* Update the separate programs/sbf/Cargo.lock for CI to pass

* Don't use jit if we're building with the sbpf-debugger feature.

* Rebase on top of master

* Update dev-bins/Cargo.lock to prevent CI from failing