cargo-install-all: modularise the binary selection (#9515)
cargo-install-all: modularise the binary selection, document script arguments

Not all arguments to the cargo-install-all scripts have been properly documented in the usage,
this has been fixed. Also the --validator-only flag has been deprecated, as the name implies
that only validator binaries are being built, which is false, as regular end-user binaries
are built as well.

Now all different kinds of binaries can be un-selected, as well as the perf-libs.

Co-authored-by: Mircea Colonescu <me_mircea@hotmail.com>