build-sbf: Fix syscalls.txt, inline them (#9346)
* build-sbf: Fix syscalls.txt, inline them

#### Problem

The syscalls generation script is broken, and isn't actually generating
anything in syscalls.txt.

Also, syscalls.txt is currently used from the SBF SDK directory, but we
could simply embed them in cargo-build-sbf directly.

#### Summary of changes

First, fix the regex in gen-syscall-list so that it covers normal and
feature-gated syscalls.

Next, make gen-syscall-list also generate a syscalls.rs file for
cargo-build-sbf, and use that directly in cargo-build-sbf. With #9345,
there won't be any more usage of SBF SDK in cargo-build-sbf.

* Print an error if something goes wrong, add comment

* Be more defensive, syscalls.txt might not exist

* Read old number of syscalls from rust file instead